refactor: components 폴더 제거
