+++ frontend/src/pages/dashboard/ui/DashboardErrorState.tsx
