+++ frontend/src/pages/dashboard/ui/MonthlyTrendSection.tsx
