+++ frontend/src/pages/dashboard/ui/TypeBreakdownSection.tsx
